Initial Upload
This is the starting of a Bootstrap 4 based website that uses simple PHP-PDO SQL with a generic login system. The system isn't yet completed, but will be updated over time as I complete this project. I work on this in my spare time, so updates will happen, I just don't know when.